+++ mongo_db_data.py
@@ -28,26 +28,41 @@ exit 2
 
     Usage:
         mongo_db_data.py -c cfg_file -d path
-            {-I -b db_name -t coll_name -a name -f {path/file | path/file*} |
+            {-I -b db_name -t coll_name -f {path/file | path/file*} [-a name]
+                [-m path] [-r] |
+            {-K -b db_name -t coll_name -f {path/file | path/file*} [-a name]
+                [-m path] [-r] |
              -D -b db_name -t coll_name
                 {-k1 "key" -l1 "value1" ["value2" "value3" ...]
                  [-k[2-5] "key" -l[2-5] "value1" ["value2" "value3" ...]] |
                  [-f {path/file | path/file*}] [-a name]} |
              -T -b db_name -t coll_name [-a name]}
-             [-p path}
+            [-p path]
             [-v | -h]
 
     Arguments:
-        -c cfg_file => Mongo configuration file.  Required arg.
-        -d dir path => Config directory path.  Required arg.
+        -c cfg_file => Mongo configuration file.
+        -d dir path => Config directory path.
 
-        -I => Insert JSON document into database.
+        -I => Import JSON document into database.
+            -b db_name => Database Name.
+            -t coll_name => Collection Name.
+            -f file(s) => JSON document to be imported.  Requires absolute
+                path.
+            -a name => Authentication Database Name.  Required for accounts
+                not in database (-b option).
+            -m path => Archive the file(s) to the directory.
+            -r => Remove the files once the insertion is completed.
+
+        -K => Insert JSON document entries into database.
             -b db_name => Database Name.
             -t coll_name => Collection Name.
             -f file(s) => JSON document to be inserted.  Requires absolute
                 path.
             -a name => Authentication Database Name.  Required for accounts
                 not in database (-b option).
+            -m path => Archive the file(s) to the directory.
+            -r => Remove the files once the insertion is completed.
 
         -D => Delete JSON document from database.
             -b db_name => Database Name.
@@ -72,9 +87,13 @@ exit 2
         -h => Help and usage message.
 
         NOTE 1:  -v and/or -h overrides all other options.
-        NOTE 2:  -I and -D are XOR options.
+        NOTE 2:  -I, -K, -T and -D are XOR options.
         NOTE 3:  -k1 and -l1 are required to be paried for -D option.
             -k[2-5] and -l[2-5] are optional, but are required to be paired.
+        NOTE 4:  The difference between -I and -K options are: -I uses the
+            mongoimport command to insert in bulk all entries into a
+            collection, whereas the -K option uses the insert command to insert
+            one entry at a time into the collection.
 
     Notes:
         Mongo configuration file format (config/mongo.py.TEMPLATE).  The
@@ -126,8 +145,8 @@ exit 2
                     tls_certkey = None
                     tls_certkey_phrase = None
 
-            Note:  FIPS Environment for Mongo.
-              If operating in a FIPS 104-2 environment, this package will
+            Note:  Secure Environment for Mongo.
+              If operating in a secure environment, this package will
               require at least a minimum of pymongo==3.8.0 or better.  It will
               also require a manual change to the auth.py module in the pymongo
               package.  See below for changes to auth.py.
@@ -155,7 +174,9 @@ exit 2
 
 # Standard
 import sys
+import os
 import subprocess
+import json
 
 # Local
 try:
@@ -189,96 +210,125 @@ def help_message():
     print(__doc__)
 
 
-def get_repset_name(svr_cfg):
+def is_file_deletable(fname):
 
-    """Function:  get_repset_name
+    """Function:  is_file_deletable
 
-    Description:  Fetch the Replication Set Name from the condfiguration file
-        or from the Mongo database.
+    Description:  Determine if a file can be removed.
 
     Arguments:
-        (input) svr_cfg -> Server configuration module
-        (output) rep_set -> Replication set name
+        (input) fname -> Filename and full path
+        (output) status -> True|False - File can be removed
 
     """
 
-    # Only pass authorization mechanism if present.
-    auth_mech = {"auth_mech": svr_cfg.auth_mech} if hasattr(
-        svr_cfg, "auth_mech") else {}
+    status = False
+    f_dirname = os.path.dirname(fname)
 
-    try:
-        rep_set = svr_cfg.repset
+    # Ensure object is a file and writable and the directory has correct perms
+    if os.path.isfile(fname):
+        if os.access(fname, os.W_OK):
+            if os.access(f_dirname, os.W_OK | os.X_OK):
+                status = True
 
-    except AttributeError:
-        coll = mongo_class.Coll(
-            svr_cfg.name, svr_cfg.user, svr_cfg.japd, host=svr_cfg.host,
-            port=svr_cfg.port, db="local", coll="system.replset",
-            auth=svr_cfg.auth, conf_file=svr_cfg.conf_file,
-            ssl_client_ca=svr_cfg.ssl_client_ca,
-            ssl_client_cert=svr_cfg.ssl_client_cert,
-            ssl_client_key=svr_cfg.ssl_client_key,
-            ssl_client_phrase=svr_cfg.ssl_client_phrase,
-            auth_type=svr_cfg.auth_type, tls_ca_certs=svr_cfg.tls_ca_certs,
-            tls_certkey=svr_cfg.tls_certkey,
-            tls_certkey_phrase=svr_cfg.tls_certkey_phrase, **auth_mech)
-        status = coll.connect()
-        rep_set = None
-
-        if status[0]:
-            if coll.coll_cnt() != 0:
-                rep_set = coll.coll_find1().get("_id")
-
-            mongo_libs.disconnect([coll])
-
-        else:
-            print(f"get_repset_name: Connection failure:  {status[1]}")
-
-    return rep_set
+    return status
 
 
-def get_repset_hosts(svr_cfg):
+def post_process(args):
 
-    """Function:  get_repset_hosts
+    """Function:  post_process
 
-    Description:  See if the Rep Set hosts is in the config file, otherwise set
-        to None.
+    Description:  Post processing of files.
 
     Arguments:
-        (input) svr_cfg -> Server Configuration module
-        (output) repset_hosts -> Contain string of rep set hosts
+        (input) args -> ArgParser class instance
 
     """
 
-    try:
-        repset_hosts = svr_cfg.repset_hosts
+    if args.arg_exist("-m"):
 
-    except AttributeError:
-        repset_hosts = None
+        for fname in args.get_val("-f"):
 
-    return repset_hosts
+            # Check if file is removable and archive directory writable
+            if args.arg_dir_chk(dir_perms_chk={"-m": 7}) \
+               and is_file_deletable(fname):
+                gen_libs.mv_file2(fname, args.get_val("-m"))
+
+            else:
+                print("post_process: Incorrect perms for file or directory")
+                print(f'\tDirectory: {args.get_val("-m")}')
+                print(f"\tFile: {fname}")                
+
+    elif args.arg_exist("-r"):
+        for fname in args.get_val("-f"):
+            status = gen_libs.rm_file(fname)
+
+            if status[0]:
+                print("post_process: Error encountered during file removal")
+                print(f"Error Message: {status[1]}")
 
 
-def insert_doc(repclu, args, **kwargs):
+def insert_doc2(coll, args, **kwargs):
 
-    """Function:  insert_doc
+    """Function:  insert_doc2
 
-    Description:  Insertion of document(s) into a Mongo database.  The document
-        can either be an external JSON document(s) or created internally
-        and then inserted into the database.  Have the ability to
-        insert multiple external JSON document files into the database.
+    Description:  Insert of document(s) into a Mongo database.  Have the
+        ability to insert multiple external JSON document files into the
+        database.
 
     Arguments:
-        (input) repclu -> Replication set/cluster instance
+        (input) coll -> Mongo collection instance
         (input) args -> ArgParser class instance
         (input) kwargs:
+            cfg -> Configuration setup
             opt_arg -> Contains list of optional arguments for command line
             opt_rep -> Contains list of replaceable arguments for command line
 
     """
 
+    for fname in args.get_val("-f"):
+        with open(fname, mode="r", encoding="UTF-8") as fhdr:
+
+            for data in fhdr:
+                status = mongo_libs.ins_doc(
+                    kwargs.get("cfg"), args.get_val("-b"), args.get_val("-t"),
+                    json.loads(data))
+
+                if not status[0]:
+                    print("insert_doc2: Insertion into Mongo failed.")
+                    print(f"Mongo error message:  {status[1]}")
+                    print(f"Data: {data}")
+
+    if args.arg_exist("-m") or args.arg_exist("-r"):
+        post_process(args)
+
+
+def insert_doc(coll, args, **kwargs):
+
+    """Function:  insert_doc
+
+    Description:  Import of document(s) into a Mongo database.  Have the
+        ability to import multiple external JSON document files into the
+        database.
+
+    Arguments:
+        (input) coll -> Mongo collection instance
+        (input) args -> ArgParser class instance
+        (input) kwargs:
+            cfg -> Configuration setup
+            opt_arg -> Contains list of optional arguments for command line
+            opt_rep -> Contains list of replaceable arguments for command line
+
+    """
+
+    use_repset = False
+
+    if isinstance(coll, mongo_class.RepSetColl):
+        use_repset = True
+
     if args.arg_exist("-f"):
         cmd = mongo_libs.create_cmd(
-            repclu, args, "mongoimport", "-p", use_repset=True, **kwargs)
+            coll, args, "mongoimport", "-p", use_repset=use_repset, **kwargs)
         orig_cmd = list(cmd)
 
         # Process files and add --file option
@@ -288,6 +338,9 @@ def insert_doc(repclu, args, **kwargs):
             proc1 = subprocess.Popen(upd_cmd)           # pylint:disable=R1732
             proc1.wait()
             cmd = list(orig_cmd)
+
+    if args.arg_exist("-m") or args.arg_exist("-r"):
+        post_process(args)
 
 
 def process_args(args):
@@ -334,7 +387,7 @@ def process_args(args):
     return status, qry
 
 
-def delete_docs(repclu, args, **kwargs):                # pylint:disable=W0613
+def delete_docs(coll, args, **kwargs):                # pylint:disable=W0613
 
     """Function:  delete_docs
 
@@ -344,89 +397,49 @@ def delete_docs(repclu, args, **kwargs):                # pylint:disable=W0613
         set of value(s).
 
     Arguments:
-        (input) repclu -> Replication set/cluster instance
+        (input) coll -> Mongo collection instance
         (input) args -> ArgParser class instance
         (input) kwargs:
+            cfg -> Configuration setup
             opt_arg -> Contains list of optional arguments for command line
             opt_rep -> Contains list of replaceable arguments for command line
 
     """
 
-    coll = mongo_class.RepSetColl(
-        repclu.name, repclu.user, repclu.japd, host=repclu.host,
-        port=repclu.port, auth=repclu.auth, repset=repclu.repset,
-        repset_hosts=repclu.repset_hosts, db=args.get_val("-b"),
-        auth_db=args.get_val("-a", def_val=repclu.auth_db),
-        coll=args.get_val("-t"), auth_mech=repclu.auth_mech,
-        ssl_client_ca=repclu.ssl_client_ca,
-        ssl_client_cert=repclu.ssl_client_cert,
-        ssl_client_key=repclu.ssl_client_key,
-        ssl_client_phrase=repclu.ssl_client_phrase,
-        auth_type=repclu.auth_type, tls_ca_certs=repclu.tls_ca_certs,
-        tls_certkey=repclu.tls_certkey,
-        tls_certkey_phrase=repclu.tls_certkey_phrase)
-    status = coll.connect()
+    if args.arg_exist("-f"):
 
-    if status[0]:
-        if args.arg_exist("-f"):
+        for fname in args.get_val("-f"):
+            lines = gen_libs.file_2_list(fname)
 
-            for fname in args.get_val("-f"):
-                lines = gen_libs.file_2_list(fname)
+            # Process each line as a delete
+            for qry in lines:
+                coll.coll_del_many(gen_libs.str_2_type(qry))
 
-                # Process each line as a delete
-                for qry in lines:
-                    coll.coll_del_many(gen_libs.str_2_type(qry))
-
-        # Assume -kN and -lN options
-        else:
-            status, qry = process_args(args)
-
-            if not status:
-                coll.coll_del_many(qry)
-
-        mongo_libs.disconnect([coll])
-
+    # Assume -kN and -lN options
     else:
-        print(f"delete_docs: Connection failure:  {status[1]}")
+        status, qry = process_args(args)
+
+        if not status:
+            coll.coll_del_many(qry)
 
 
-def truncate_coll(repclu, args, **kwargs):              # pylint:disable=W0613
+def truncate_coll(coll, args, **kwargs):
 
     """Function:  truncate_coll
 
     Description:  Truncate a collection in a Mongo database.
 
     Arguments:
-        (input) repclu -> Replication set/cluster instance
+        (input) coll -> Mongo collection instance
         (input) args -> ArgParser class instance
         (input) kwargs:
+            cfg -> Configuration setup
             opt_arg -> Contains list of optional arguments for command line
             opt_rep -> Contains list of replaceable arguments for command line
 
     """
 
-    coll = mongo_class.RepSetColl(
-        repclu.name, repclu.user, repclu.japd, host=repclu.host,
-        port=repclu.port, auth=repclu.auth, repset=repclu.repset,
-        repset_hosts=repclu.repset_hosts, db=args.get_val("-b"),
-        coll=args.get_val("-t"), auth_mech=repclu.auth_mech,
-        auth_db=args.get_val("-a", def_val=repclu.auth_db),
-        ssl_client_ca=repclu.ssl_client_ca,
-        ssl_client_cert=repclu.ssl_client_cert,
-        ssl_client_key=repclu.ssl_client_key,
-        ssl_client_phrase=repclu.ssl_client_phrase,
-        auth_type=repclu.auth_type, tls_ca_certs=repclu.tls_ca_certs,
-        tls_certkey=repclu.tls_certkey,
-        tls_certkey_phrase=repclu.tls_certkey_phrase)
-    status = coll.connect()
-
-    if status[0]:
-        # Require override option
-        coll.coll_del_many({}, True)
-        mongo_libs.disconnect([coll])
-
-    else:
-        print(f"truncate_coll: Connection failure:  {status[1]}")
+    coll.coll_del_many({}, True)
 
 
 def run_program(args, func_dict, **kwargs):
@@ -445,34 +458,28 @@ def run_program(args, func_dict, **kwargs):
     """
 
     func_dict = dict(func_dict)
-    svr_cfg = gen_libs.load_module(args.get_val("-c"), args.get_val("-d"))
-    rep_set = get_repset_name(svr_cfg)
-    repset_hosts = get_repset_hosts(svr_cfg)
+    cfg = gen_libs.load_module(args.get_val("-c"), args.get_val("-d"))
 
-    # Only pass authorization mechanism if present
-    auth_mech = {"auth_mech": svr_cfg.auth_mech} if hasattr(
-        svr_cfg, "auth_mech") else {}
+    if cfg.repset:
+        coll = mongo_libs.create_instance(
+            args.get_val("-c"), args.get_val("-d"), mongo_class.RepSetColl)
+        coll.db_name = args.get_val("-b")
+        coll.coll = args.get_val("-t")
 
-    repclu = mongo_class.RepSet(
-        svr_cfg.name, svr_cfg.user, svr_cfg.japd, host=svr_cfg.host,
-        port=svr_cfg.port, auth=svr_cfg.auth, repset=rep_set,
-        repset_hosts=repset_hosts,
-        auth_db=args.get_val("-a", def_val=svr_cfg.auth_db),
-        ssl_client_ca=svr_cfg.ssl_client_ca,
-        ssl_client_cert=svr_cfg.ssl_client_cert,
-        ssl_client_key=svr_cfg.ssl_client_key,
-        ssl_client_phrase=svr_cfg.ssl_client_phrase,
-        auth_type=svr_cfg.auth_type, tls_ca_certs=svr_cfg.tls_ca_certs,
-        tls_certkey=svr_cfg.tls_certkey,
-        tls_certkey_phrase=svr_cfg.tls_certkey_phrase, **auth_mech)
-    status = repclu.connect()
+    else:
+        coll = mongo_libs.create_instance(
+            args.get_val("-c"), args.get_val("-d"), mongo_class.Coll)
+        coll.coll_db = args.get_val("-b")
+        coll.coll_coll = args.get_val("-t")
+
+    status = coll.connect()
 
     if status[0]:
         # Intersect args_array & func_dict to determine which functions to call
         for func in set(args.get_args_keys()) & set(func_dict.keys()):
-            func_dict[func](repclu, args, **kwargs)
+            func_dict[func](coll, args, cfg=cfg, **kwargs)
 
-        mongo_libs.disconnect([repclu])
+        mongo_libs.disconnect([coll])
 
     else:
         print(f"run_program: Connection failure:  {status[1]}")
@@ -505,35 +512,38 @@ def main():
 
     dir_perms_chk = {"-d": 5, "-p": 5}
     file_perm_chk = {"-f": 6}
-    func_dict = {"-I": insert_doc, "-D": delete_docs, "-T": truncate_coll}
+    func_dict = {
+        "-I": insert_doc, "-D": delete_docs, "-T": truncate_coll,
+        "-K": insert_doc2}
     opt_arg_list = {
         "-a": "--authenticationDatabase=", "-b": "--db=",
         "-t": "--collection="}
     opt_arg_rep = {"-f": "--file="}
-    opt_con_req_list = {"-k1": ["-l1"]}
+    opt_con_req_list = {
+        "-k1": ["-l1"], "-k2": ["-l2"], "-k3": ["-l3"], "-k4": ["-l4"],
+        "-k5": ["-l5"], "-I": ["-b", "-t", "-f"], "-K": ["-b", "-t", "-f"],
+        "-D": ["-b", "-t"], "-T": ["-b", "-t"]}
     opt_multi_list = ["-f", "-l1", "-l2", "-l3", "-l4", "-l5"]
     opt_req_list = ["-b", "-c", "-d", "-t"]
     opt_val_list = [
         "-a", "-b", "-c", "-d", "-f", "-k1", "-l1", "-p", "-t", "-k2", "-l2",
-        "-k3", "-l3", "-k4", "-l4", "-k5", "-l5"]
-    opt_xor_dict = {"-D": ["-I", "-T"], "-I": ["-D", "-T"], "-T": ["-D", "-I"]}
+        "-k3", "-l3", "-k4", "-l4", "-k5", "-l5", "-m"]
+    opt_xor_dict = {
+        "-D": ["-I", "-T", "-K"], "-I": ["-D", "-T", "-K"],
+        "-T": ["-D", "-I", "-K"], "-K": ["-D", "-T", "-I"]}
     xor_noreq_list = {"-k1": "-f"}
 
     # Process argument list from command line
     args = gen_class.ArgParser(
-        sys.argv, opt_val=opt_val_list, multi_val=opt_multi_list,
-        do_parse=True)
+        sys.argv, opt_val=opt_val_list, multi_val=opt_multi_list)
 
-    # Remove dupe files
-    if args.arg_exist("-f"):
-        args.update_arg("-f", gen_libs.rm_dup_list(args.get_val("-f")))
-
-    if not gen_libs.help_func(args, __version__, help_message)  \
-       and args.arg_require(opt_req=opt_req_list)               \
-       and args.arg_dir_chk(dir_perms_chk=dir_perms_chk)        \
-       and args.arg_xor_dict(opt_xor_val=opt_xor_dict)          \
-       and args.arg_cond_req(opt_con_req=opt_con_req_list)      \
-       and args.arg_noreq_xor(xor_noreq=xor_noreq_list)         \
+    if args.arg_parse2()                                            \
+       and not gen_libs.help_func(args, __version__, help_message)  \
+       and args.arg_require(opt_req=opt_req_list)                   \
+       and args.arg_dir_chk(dir_perms_chk=dir_perms_chk)            \
+       and args.arg_xor_dict(opt_xor_val=opt_xor_dict)              \
+       and args.arg_cond_req(opt_con_req=opt_con_req_list)          \
+       and args.arg_noreq_xor(xor_noreq=xor_noreq_list)             \
        and args.arg_file_chk(file_perm_chk=file_perm_chk):
         run_program(
             args, func_dict, opt_arg=opt_arg_list, opt_rep=opt_arg_rep)