+++ test/unit/mongo_db_data/delete_docs.py
@@ -78,15 +78,14 @@ class ArgParser():
         return arg in self.args_array
 
 
-class RepSetColl():
+class Coll():                                           # pylint:disable=R0903
 
-    """Class:  RepSetColl
+    """Class:  Coll
 
-    Description:  Class stub holder for mongo_class.RepSetColl class.
+    Description:  Class stub holder for mongo_class.Coll class.
 
     Methods:
         __init__
-        connect
         coll_del_many
 
     """
@@ -102,20 +101,6 @@ class RepSetColl():
         """
 
         self.query = None
-        self.status = True
-        self.err_msg = None
-
-    def connect(self):
-
-        """Method:  connect
-
-        Description:  Stub holder for mongo_class.RepSetColl.connect method.
-
-        Arguments:
-
-        """
-
-        return self.status, self.err_msg
 
     def coll_del_many(self, query):
 
@@ -131,48 +116,6 @@ class RepSetColl():
         self.query = query
 
 
-class RepSetCfg():                                      # pylint:disable=R0903
-
-    """Class:  RepSetCfg
-
-    Description:  Class which is a representation of a RepSet class.
-
-    Methods:
-        __init__
-
-    """
-
-    def __init__(self):
-
-        """Method:  __init__
-
-        Description:  Initialization instance of the CfgTest class.
-
-        Arguments:
-
-        """
-
-        self.name = "MongoName"
-        self.user = "root"
-        self.japd = None
-        self.host = "HostName"
-        self.port = 27017
-        self.auth = True
-        self.conf_file = "ConFile"
-        self.repset = "RepSetName"
-        self.repset_hosts = ["List of hosts"]
-        self.auth_mech = "SCRAM-SHA-1"
-        self.auth_db = "admin"
-        self.ssl_client_ca = None
-        self.ssl_client_cert = None
-        self.ssl_client_key = None
-        self.ssl_client_phrase = None
-        self.auth_type = None
-        self.tls_ca_certs = None
-        self.tls_certkey = None
-        self.tls_certkey_phrase = None
-
-
 class UnitTest(unittest.TestCase):
 
     """Class:  UnitTest
@@ -181,10 +124,6 @@ class UnitTest(unittest.TestCase):
 
     Methods:
         setUp
-        test_without_a_option
-        test_with_a_option
-        test_connection_fail
-        test_connection_success
         test_multiple_lines
         test_multiple_files
         test_empty_file
@@ -204,8 +143,7 @@ class UnitTest(unittest.TestCase):
 
         """
 
-        self.repset = RepSetCfg()
-        self.repcoll = RepSetColl()
+        self.coll = Coll()
         self.args = ArgParser()
         self.args2 = ArgParser()
         self.args3 = ArgParser()
@@ -221,90 +159,9 @@ class UnitTest(unittest.TestCase):
         self.args4.args_array = {
             "-b": "databasename", "-t": "tablename", "-f": ["file1"]}
 
-    @mock.patch("mongo_db_data.mongo_libs.disconnect")
     @mock.patch("mongo_db_data.gen_libs")
     @mock.patch("mongo_db_data.mongo_class.RepSetColl")
-    def test_without_a_option(self, mock_coll, mock_lib, mock_disconnect):
-
-        """Function:  test_without_a_option
-
-        Description:  Test without the -a option.
-
-        Arguments:
-
-        """
-
-        mock_coll.return_value = self.repcoll
-        mock_lib.file_2_list.return_value = ["File1"]
-        mock_lib.str_2_type.return_value = {"query"}
-        mock_disconnect.return_value = True
-
-        self.assertFalse(mongo_db_data.delete_docs(self.repset, self.args4))
-
-    @mock.patch("mongo_db_data.mongo_libs.disconnect")
-    @mock.patch("mongo_db_data.gen_libs")
-    @mock.patch("mongo_db_data.mongo_class.RepSetColl")
-    def test_with_a_option(self, mock_coll, mock_lib, mock_disconnect):
-
-        """Function:  test_with_a_option
-
-        Description:  Test with the -a option.
-
-        Arguments:
-
-        """
-
-        mock_coll.return_value = self.repcoll
-        mock_lib.file_2_list.return_value = ["File1"]
-        mock_lib.str_2_type.return_value = {"query"}
-        mock_disconnect.return_value = True
-
-        self.assertFalse(mongo_db_data.delete_docs(self.repset, self.args2))
-
-    @mock.patch("mongo_db_data.mongo_class.RepSetColl")
-    def test_connection_fail(self, mock_coll):
-
-        """Function:  test_connection_fail
-
-        Description:  Test with failed connection.
-
-        Arguments:
-
-        """
-
-        self.repcoll.status = False
-        self.repcoll.err_msg = "Error Connection Message"
-
-        mock_coll.return_value = self.repcoll
-
-        with gen_libs.no_std_out():
-            self.assertFalse(
-                mongo_db_data.delete_docs(self.repset, self.args2))
-
-    @mock.patch("mongo_db_data.mongo_libs.disconnect")
-    @mock.patch("mongo_db_data.gen_libs")
-    @mock.patch("mongo_db_data.mongo_class.RepSetColl")
-    def test_connection_success(self, mock_coll, mock_lib, mock_disconnect):
-
-        """Function:  test_connection_success
-
-        Description:  Test with successful connection.
-
-        Arguments:
-
-        """
-
-        mock_coll.return_value = self.repcoll
-        mock_lib.file_2_list.return_value = ["File1"]
-        mock_lib.str_2_type.return_value = {"query"}
-        mock_disconnect.return_value = True
-
-        self.assertFalse(mongo_db_data.delete_docs(self.repset, self.args2))
-
-    @mock.patch("mongo_db_data.mongo_libs.disconnect")
-    @mock.patch("mongo_db_data.gen_libs")
-    @mock.patch("mongo_db_data.mongo_class.RepSetColl")
-    def test_multiple_lines(self, mock_coll, mock_lib, mock_disconnect):
+    def test_multiple_lines(self, mock_coll, mock_lib):
 
         """Function:  test_multiple_lines
 
@@ -314,17 +171,15 @@ class UnitTest(unittest.TestCase):
 
         """
 
-        mock_coll.return_value = self.repcoll
+        mock_coll.return_value = self.coll
         mock_lib.file_2_list.return_value = ["file1", "file2"]
         mock_lib.str_2_type.return_value = {"query"}
-        mock_disconnect.return_value = True
 
-        self.assertFalse(mongo_db_data.delete_docs(self.repset, self.args2))
+        self.assertFalse(mongo_db_data.delete_docs(self.coll, self.args2))
 
-    @mock.patch("mongo_db_data.mongo_libs.disconnect")
     @mock.patch("mongo_db_data.gen_libs")
     @mock.patch("mongo_db_data.mongo_class.RepSetColl")
-    def test_multiple_files(self, mock_coll, mock_lib, mock_disconnect):
+    def test_multiple_files(self, mock_coll, mock_lib):
 
         """Function:  test_multiple_files
 
@@ -334,17 +189,15 @@ class UnitTest(unittest.TestCase):
 
         """
 
-        mock_coll.return_value = self.repcoll
+        mock_coll.return_value = self.coll
         mock_lib.file_2_list.return_value = ["file1", "file2"]
         mock_lib.str_2_type.return_value = {"query"}
-        mock_disconnect.return_value = True
 
-        self.assertFalse(mongo_db_data.delete_docs(self.repset, self.args3))
+        self.assertFalse(mongo_db_data.delete_docs(self.coll, self.args3))
 
-    @mock.patch("mongo_db_data.mongo_libs.disconnect")
     @mock.patch("mongo_db_data.gen_libs")
     @mock.patch("mongo_db_data.mongo_class.RepSetColl")
-    def test_empty_file(self, mock_coll, mock_lib, mock_disconnect):
+    def test_empty_file(self, mock_coll, mock_lib):
 
         """Function:  test_empty_file
 
@@ -354,16 +207,14 @@ class UnitTest(unittest.TestCase):
 
         """
 
-        mock_coll.return_value = self.repcoll
+        mock_coll.return_value = self.coll
         mock_lib.file_2_list.return_value = []
-        mock_disconnect.return_value = True
 
-        self.assertFalse(mongo_db_data.delete_docs(self.repset, self.args2))
+        self.assertFalse(mongo_db_data.delete_docs(self.coll, self.args2))
 
-    @mock.patch("mongo_db_data.mongo_libs.disconnect")
     @mock.patch("mongo_db_data.gen_libs")
     @mock.patch("mongo_db_data.mongo_class.RepSetColl")
-    def test_file_list(self, mock_coll, mock_lib, mock_disconnect):
+    def test_file_list(self, mock_coll, mock_lib):
 
         """Function:  test_file_list
 
@@ -373,17 +224,15 @@ class UnitTest(unittest.TestCase):
 
         """
 
-        mock_coll.return_value = self.repcoll
+        mock_coll.return_value = self.coll
         mock_lib.file_2_list.return_value = ["File1"]
         mock_lib.str_2_type.return_value = {"query"}
-        mock_disconnect.return_value = True
 
-        self.assertFalse(mongo_db_data.delete_docs(self.repset, self.args2))
+        self.assertFalse(mongo_db_data.delete_docs(self.coll, self.args2))
 
-    @mock.patch("mongo_db_data.mongo_libs.disconnect")
     @mock.patch("mongo_db_data.process_args")
     @mock.patch("mongo_db_data.mongo_class.RepSetColl")
-    def test_no_list_error(self, mock_coll, mock_proc, mock_disconnect):
+    def test_no_list_error(self, mock_coll, mock_proc):
 
         """Function:  test_no_list_error
 
@@ -393,16 +242,14 @@ class UnitTest(unittest.TestCase):
 
         """
 
-        mock_coll.return_value = self.repcoll
+        mock_coll.return_value = self.coll
         mock_proc.return_value = (False, {"Query"})
-        mock_disconnect.return_value = True
 
-        self.assertFalse(mongo_db_data.delete_docs(self.repset, self.args))
+        self.assertFalse(mongo_db_data.delete_docs(self.coll, self.args))
 
-    @mock.patch("mongo_db_data.mongo_libs.disconnect")
     @mock.patch("mongo_db_data.process_args")
     @mock.patch("mongo_db_data.mongo_class.RepSetColl")
-    def test_no_file_list(self, mock_coll, mock_proc, mock_disconnect):
+    def test_no_file_list(self, mock_coll, mock_proc):
 
         """Function:  test_no_file_list
 
@@ -412,11 +259,10 @@ class UnitTest(unittest.TestCase):
 
         """
 
-        mock_coll.return_value = self.repcoll
+        mock_coll.return_value = self.coll
         mock_proc.return_value = (False, {"Query"})
-        mock_disconnect.return_value = True
 
-        self.assertFalse(mongo_db_data.delete_docs(self.repset, self.args))
+        self.assertFalse(mongo_db_data.delete_docs(self.coll, self.args))
 
 
 if __name__ == "__main__":